big refactor - rethink aggregation
We currently don't aggregate properly, this attempts to fix that, or partly.

The different Quantified types have different behaviours: Ingredients
contents are fractional and can contain various Measurable types, Dish
and Meal contents are not and can only contain #Mass measurements (but
should be aggregated into a fractional list of the various basic
types of Measurables)

We need to alter the Measurables table to store the default unit type,
rather than just the dimensions. This is needed when aggregating
because the unit to use when aggregating Measurables of a content of a
Dish (for example) is otherwise not available.