various - implement NumPortions aggregation
Various API changes needed, because we can't calculate totalMass
without also making a recursive traversal of the contents (in the
general case). Each item with a units in #portions needs to have the
portion size calculated. And if that is a dish, that depends on its
total mass contents... so those need to be calculated. etc., all the
way down to basic ingredients (which have an inherent portionSize)